requirements.txt
Django==2.2.3
gunicorn==19.9.0
psycopg2-binary==2.8.3
django-crispy-forms==1.7.2
django-import-export==1.2.0